+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidytuesdayR (development version)
+
 # tidytuesdayR 1.2.1
 * [tests] No user-facing changes.